+++ packages/plugin-button/web/src/components/button-component.jsx
@@ -15,6 +15,7 @@ class ButtonComponent extends PureComponent {
     const {
       componentData: { button },
       settings: { onClick },
+      helpers,
     } = this.props;
     const { theme } = this.props;
     const buttonText = button.settings.buttonText;
@@ -38,6 +39,7 @@ class ButtonComponent extends PureComponent {
         buttonText={buttonText}
         theme={theme}
         onClick={onClick}
+        helpers={helpers}
         {...linkButtonSettings}
       />
     );
@@ -51,6 +53,7 @@ ButtonComponent.propTypes = {
   settings: PropTypes.object.isRequired,
   blockProps: PropTypes.object,
   theme: PropTypes.object.isRequired,
+  helpers: PropTypes.object,
 };
 
 export default ButtonComponent;