+++ packages/plugin-gallery/web/src/toolbar/inline-buttons.tsx
@@ -11,19 +11,20 @@ import {
 } from 'wix-rich-content-common';
 import { GalleryPluginEditorConfig } from '../types';
 
-const modalStyles = getModalStyles({});
-
 const createInlineButtons: CreateInlineButtons = ({
   t,
   anchorTarget,
   relValue,
   settings,
+  isMobile,
 }: {
   t: TranslationFunction;
   settings: GalleryPluginEditorConfig;
   anchorTarget: AnchorTarget;
   relValue: RelValue;
+  isMobile: boolean;
 }) => {
+  const modalStyles = getModalStyles({ isMobile });
   const icons = settings?.toolbar?.icons || {};
   const spoilerButton = settings.spoiler
     ? [
@@ -93,7 +94,7 @@ const createInlineButtons: CreateInlineButtons = ({
       modalStyles,
       switchLayout,
       t,
-      mobile: false,
+      mobile: true,
       tooltipTextKey: 'SettingsButton_Tooltip',
       anchorTarget,
       relValue,