+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 ### :rocket: New Feature
 - `file-upload`
   - [#2330](https://github.com/wix/ricos/commit/2330) add responsive design
+- `gallery`
+  - [#2364](https://github.com/wix/ricos/pull/2364) gallery settings for mobile devices
 
 ### :bug: Bug Fix
 - `ricos-viewer`